pre 4.0.1
copyright표기, coding방식, py파일 분할
