+++ calc_result.py
@@ -1,3 +1,7 @@
+#-*- coding: utf-8 -*-
+## 코드를 무단으로 복제하여 개조 및 배포하지 말 것##
+## Copyright ⓒ 2020 Dawnclass(새벽반) dawnclass16@naver.com
+
 result_check_list_explain=[(
 "크리티컬 확률은 장비 세팅의 유연성에 "
 "영향을 미칩니다. 직업마다 보정되는 크"
